PMM-6259 move types to correct file
