+++ pmm-app/src/pmm-qan/panel/components/Details/Details.hooks.ts
@@ -1,8 +1,7 @@
 import { useContext, useEffect, useState } from 'react';
 import { QueryAnalyticsProvider } from 'pmm-qan/panel/provider/provider';
-import { Databases } from './Details.constants';
 import DetailsService from './Details.service';
-import { DatabasesType } from './Details.types';
+import { Databases, DatabasesType } from './Details.types';
 
 export const useDetails = (): [boolean, any, DatabasesType] => {
   const {
