+++ pmm-app/src/pmm-qan/panel/components/Details/Table/TableContainer.hooks.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { get } from 'lodash';
-import { Databases } from '../Details.constants';
 import { useExplains } from '../Explain/Explain.hooks';
+import { Databases } from '../Details.types';
 
 export const useTables = (examples, databaseType): any[] => {
   const [jsonExplain, classicExplain] = useExplains(examples, databaseType);
